+++ utils/bytes.go
@@ -19,51 +19,3 @@ func ToUpperBytes(b []byte) []byte {
 	}
 	return b
 }
-
-// TrimRightBytes is the equivalent of bytes.TrimRight
-func TrimRightBytes(b []byte, cutset byte) []byte {
-	lenStr := len(b)
-	for lenStr > 0 && b[lenStr-1] == cutset {
-		lenStr--
-	}
-	return b[:lenStr]
-}
-
-// TrimLeftBytes is the equivalent of bytes.TrimLeft
-func TrimLeftBytes(b []byte, cutset byte) []byte {
-	lenStr, start := len(b), 0
-	for start < lenStr && b[start] == cutset {
-		start++
-	}
-	return b[start:]
-}
-
-// TrimBytes is the equivalent of bytes.Trim
-func TrimBytes(b []byte, cutset byte) []byte {
-	i, j := 0, len(b)-1
-	for ; i <= j; i++ {
-		if b[i] != cutset {
-			break
-		}
-	}
-	for ; i < j; j-- {
-		if b[j] != cutset {
-			break
-		}
-	}
-
-	return b[i : j+1]
-}
-
-// EqualFoldBytes tests ascii slices for equality case-insensitively
-func EqualFoldBytes(b, s []byte) bool {
-	if len(b) != len(s) {
-		return false
-	}
-	for i := len(b) - 1; i >= 0; i-- {
-		if toUpperTable[b[i]] != toUpperTable[s[i]] {
-			return false
-		}
-	}
-	return true
-}
