+++ utils/byteseq.go
@@ -1,0 +1,53 @@
+package utils
+
+type byteSeq interface {
+	~string | ~[]byte
+}
+
+// Trim is the equivalent of [strings.Trim] or [bytes.Trim].
+func Trim[S byteSeq](s S, cutset byte) S {
+	i, j := 0, len(s)-1
+	for ; i <= j; i++ {
+		if s[i] != cutset {
+			break
+		}
+	}
+	for ; i < j; j-- {
+		if s[j] != cutset {
+			break
+		}
+	}
+
+	return s[i : j+1]
+}
+
+// TrimLeft is the equivalent of [strings.TrimLeft] or [bytes.TrimLeft].
+func TrimLeft[S byteSeq](s S, cutset byte) S {
+	lenStr, start := len(s), 0
+	for start < lenStr && s[start] == cutset {
+		start++
+	}
+	return s[start:]
+}
+
+// TrimRight is the equivalent of [strings.TrimRight] or [bytes.TrimRight].
+func TrimRight[S byteSeq](s S, cutset byte) S {
+	lenStr := len(s)
+	for lenStr > 0 && s[lenStr-1] == cutset {
+		lenStr--
+	}
+	return s[:lenStr]
+}
+
+// EqualFold tests ascii strings or bytes for equality case-insensitively
+func EqualFold[S byteSeq](b, s S) bool {
+	if len(b) != len(s) {
+		return false
+	}
+	for i := len(b) - 1; i >= 0; i-- {
+		if toUpperTable[b[i]] != toUpperTable[s[i]] {
+			return false
+		}
+	}
+	return true
+}
